requirements.txt
PySide6
requests
python-nmap

